Implement Forgetful Bloom filters
Forgetful bloom filters consist of at least 3 bloom filters, future,
present and at least one past filter. When inserting into a forgetful
bloom the key is inserted into the future and present blooms. After some
rotation critera (currently just the number of insertions) the future
bloom is moved to the present bloom, the present bloom is moved to the
oldest past bloom and the oldest past bloom is cleared and moved to be
the future bloom.