Enable easier cross-compilation
When cross-compiling this NIF, the path to the built Rust library is
different than the typical `target/release/libname.so`. There are no
stable mechanisms to dictate to cargo where to output the library, but
we can infer its location in a build script. The solution is to write
the library's path to a file and use that file's contents as the
source parameter when installing the lib.